+++ packages/implementations/react/src/atomic/Label/src/index.ts
@@ -1,0 +1,5 @@
+import { Label } from "./Label";
+
+export { Label } from "./Label";
+export type { LabelProps } from "./label.types";
+export default Label;